+++ constraint/constraint.go
@@ -1,0 +1,6 @@
+// Copyright 2022 Livesport TV s.r.o. All rights reserved.
+// Use of this source code is governed by a MIT license
+// that can be found in the LICENSE file.
+
+// Package constraint contains generic support and helper functions.
+package constraint